% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stomata_anatomy}
\alias{stomata_anatomy}
\title{Stomatal anatomy}
\format{
A data frame with 5842 rows and 7 variables:
\describe{
\item{\code{source_id}}{character Taxonomic name give by the source paper.}
\item{\code{trait}}{character Description of trait.}
\item{\code{mu}}{double Trait mean.}
\item{\code{n}}{double Sample size (see \code{stomata_anatomy_metadata} for more information.)}
\item{\code{sd}}{double Trait standard deviation.}
\item{\code{source}}{character Source paper identifier.}
\item{\code{se}}{double Trait standard error.}
}
}
\usage{
stomata_anatomy
}
\description{
The data set contains stomatal anatomical characters (density, length, width, pore length, and pore width) for the ab- and adaxial leaf surface in a number of species synthesized from the primary literature.
}
\keyword{datasets}
