% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stomata_anatomy_metadata}
\alias{stomata_anatomy_metadata}
\title{Stomatal anatomy metadata}
\format{
A data frame with 38 rows and 7 variables:
\describe{
\item{\code{source}}{character Source paper identifier.}
\item{\code{taxa}}{character Description of the taxa covered in source.}
\item{\code{location}}{character Location where trait data are located in source. In some cases, raw data were provided directly by the authors.}
\item{\code{density_unit}}{character Unit used for stomatal density.}
\item{\code{length_unit}}{character Unit used for guard cell or pore length or width.}
\item{\code{sample_size}}{double Approximate sample size for each taxon in the study.}
\item{\code{notes}}{character Notes to understand units or sample size data.}
}
}
\usage{
stomata_anatomy_metadata
}
\description{
This contains metadata for stomata_anatomy
}
\keyword{datasets}
